+++ Coordinator/mlxbar/__init__.py
@@ -1,3 +1,3 @@
 """MLXBar coordinator package."""
 
-__version__ = "1.8.1"
+__version__ = "1.8.2"